Add in-memory form repository with module loader
